Support splitting a request to multiple requests
Often, there's a server side enforced limit on the number of fields sent
by a client, to protect against denial of service attacks.
By default on Django that limit is set to 1000 field values, however
that is often an issue for certain API requests.

When a single field is expected to be provided with multiple values (as
is about to happen in #329), this PR aims to provide a solution by
extending QueryWorker with the `splitable` parameter, whichh will split
that parameter to multiple requests when it exceeds a hardcoded length
of values (set to 900).

This can't be done automatically, and should be done with caution, as
only certain endpoints will yield the same result, and only with certain
parameters.

Signed-off-by: Nir Izraeli <nirizr@gmail.com>